pyproject.toml
[tool.poetry]
name = "codyv2"
version = "0.1.0"
description = ""
authors = ["musashi-miyamot0 <89585170+musashi-miyamot0@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
colorama = "^0.4.6"
loguru = "^0.7.2"
pika = "^1.3.2"
pyttsx3 = "^2.98"
sounddevice = "^0.5.1"
pydub = "^0.25.1"
grpcio-tools = ">=1.48.2"
yandex-speechkit = "^1.5.0"
torch = "^2.5.0"
vosk = "^0.3.45"
pyaudio = "^0.2.14"
google-generativeai = "^0.8.3"
soundfile = "^0.12.1"
pyyaml = "^6.0.2"
fuzzywuzzy = "^0.18.0"
python-levenshtein = "^0.26.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
